slight code adjustment
todo: create a selling logic, the npc has a coefficient to how much its willing to spend. in the future we should create a few coefficients per npc in a range of: begrudgingly willing to sell to okay with the sell, to very happy (very small range that might boost friendliness?) and if you're outside those ranges the npc gets upset and leaves, possibly hurting relations or sparking an event in the future.