Finalized dungeon workflow! Turns out there were a few small things left.
Now dungeons can be validated; You can decide when dungeons are considered value to ensure, for example, at least one shop room.
Dungeon elements now contain a path to their prefabs; This is so that the generator only loads prefabs when they are actually placed into the dungeon. Should prevent performance degradation as we incorporate more assets into the game.